# Optimize expression `LIKE() ESCAPE()` for bin collator