### What problem does this PR solve?

Issue Number: ref #5294

Problem Summary:

### What is changed and how it works?

- batch match str and optimize code

### Benchmark

ENV:
- tpch100
- 1 tiflash
- limit cpu up to 200%
- original commit: a476307afa27c5e966bc6dabb7d63f4ea93b64c1

----

SQL: subquery like tpch-q13
```
select count(1) from orders where o_comment like '%pending%deposits%';
```

Time(s) | Original | Optimized |   |   | Improvement
-- | -- | -- | -- | -- | --
  | 35.77 | 11 |   |   |  
  | 33.84 | 10.88 |   |   |  
  | 35.32 | 11.11 |   |   |  
  | 34.82 | 11.21 |   |   |  
  | 34.94 | 10.97 |   |   | AVG(Original) / AVG(Optimized) - 1.0
AVG | 34.938 | 11.034 |   | Optimized : Original | 216.64%

### TODO
- seachers like `ASCIICaseSensitiveStringSearcher` or `Volnitsky` are too heavy for small str
- Aho–Corasick algorithm is NOT friendly for cache locality.
- Need to optimize `strstr` search by simd in next pull request.

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```